[FIX] Other cl warnings (signed/unsigned mismatch warning).
git-svn-id: https://b-tk.googlecode.com/svn/BTK/trunk@347 93be6e1e-6fcc-11de-baec-69240fdc299a

+++ Code/BasicFilters/btkAcquisitionUnitConverter.h
@@ -98,9 +98,9 @@ namespace btk
       return false;
     else
     {
-      for (int i = 0 ; i < cols ; ++i)
+      for (size_t i = 0 ; i < cols ; ++i)
       {
-        for (int j = 0 ; j < rows ; ++j)
+        for (size_t j = 0 ; j < rows ; ++j)
         {
           values[j + i*columnsStep + alreadyConverted] = static_cast<float>(values[j + i*columnsStep + alreadyConverted] * scales[j]);
         }
